pyproject.toml
[tool.poetry]
name = "grbl2"
version = "0.1.0"
description = "Mirai-QQ-Robot"
authors = ["shitlime"]

[tool.poetry.dependencies]
python = "^3.10"
graia-ariadne = {extras = ["standard"], version = "^0.7.16"}
pypinyin = "^0.46.0"
pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
# 这里以清华源举例，你也可以使用其他源
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false
